Moves actions and datasources to subpackages
