+++ SeviBus/src/main/java/com/sloy/sevibus/resources/StuffProvider.java
@@ -7,24 +7,24 @@ import com.firebase.client.Firebase;
 import com.j256.ormlite.android.apptools.OpenHelperManager;
 import com.sloy.sevibus.BuildConfig;
 import com.sloy.sevibus.bbdd.DBHelper;
-import com.sloy.sevibus.resources.actions.DeleteFavoritaAction;
 import com.sloy.sevibus.resources.actions.ObtainCercanasAction;
-import com.sloy.sevibus.resources.actions.ObtainFavoritasAction;
 import com.sloy.sevibus.resources.actions.ObtainLineasCercanasAction;
-import com.sloy.sevibus.resources.actions.ObtainLlegadasAction;
-import com.sloy.sevibus.resources.actions.ObtainSingleFavoritaAction;
-import com.sloy.sevibus.resources.actions.ReorderFavoritasAction;
-import com.sloy.sevibus.resources.actions.SaveFavoritaAction;
+import com.sloy.sevibus.resources.actions.favorita.DeleteFavoritaAction;
+import com.sloy.sevibus.resources.actions.favorita.ObtainFavoritasAction;
+import com.sloy.sevibus.resources.actions.llegada.ObtainLlegadasAction;
+import com.sloy.sevibus.resources.actions.favorita.ObtainSingleFavoritaAction;
+import com.sloy.sevibus.resources.actions.favorita.ReorderFavoritasAction;
+import com.sloy.sevibus.resources.actions.favorita.SaveFavoritaAction;
 import com.sloy.sevibus.resources.datasource.ApiErrorHandler;
-import com.sloy.sevibus.resources.datasource.ApiLlegadaDataSource;
-import com.sloy.sevibus.resources.datasource.DBFavoritaDataSource;
-import com.sloy.sevibus.resources.datasource.FavoritaDataSource;
 import com.sloy.sevibus.resources.datasource.LineaDataSource;
-import com.sloy.sevibus.resources.datasource.FirebaseFavoritaDataSource;
-import com.sloy.sevibus.resources.datasource.LlegadaDataSource;
-import com.sloy.sevibus.resources.datasource.SevibusApi;
+import com.sloy.sevibus.resources.datasource.llegada.ApiLlegadaDataSource;
+import com.sloy.sevibus.resources.datasource.favorita.DBFavoritaDataSource;
+import com.sloy.sevibus.resources.datasource.favorita.FavoritaDataSource;
+import com.sloy.sevibus.resources.datasource.favorita.FirebaseFavoritaDataSource;
+import com.sloy.sevibus.resources.datasource.llegada.LlegadaDataSource;
+import com.sloy.sevibus.resources.datasource.llegada.SevibusApi;
 import com.sloy.sevibus.resources.datasource.StringDownloader;
-import com.sloy.sevibus.resources.datasource.TussamLlegadaDataSource;
+import com.sloy.sevibus.resources.datasource.llegada.TussamLlegadaDataSource;
 import com.sloy.sevibus.resources.sync.UpdateDatabaseAction;
 
 import retrofit.RestAdapter;
