+++ SeviBus/src/main/java/com/sloy/sevibus/resources/actions/favorita/DeleteFavoritaAction.java
@@ -1,6 +1,6 @@
-package com.sloy.sevibus.resources.actions;
+package com.sloy.sevibus.resources.actions.favorita;
 
-import com.sloy.sevibus.resources.datasource.FavoritaDataSource;
+import com.sloy.sevibus.resources.datasource.favorita.FavoritaDataSource;
 
 import rx.Observable;
