+++ SeviBus/src/main/java/com/sloy/sevibus/resources/actions/favorita/ObtainFavoritasAction.java
@@ -1,7 +1,7 @@
-package com.sloy.sevibus.resources.actions;
+package com.sloy.sevibus.resources.actions.favorita;
 
 import com.sloy.sevibus.model.tussam.Favorita;
-import com.sloy.sevibus.resources.datasource.FavoritaDataSource;
+import com.sloy.sevibus.resources.datasource.favorita.FavoritaDataSource;
 
 import java.util.List;
