+++ SeviBus/src/main/java/com/sloy/sevibus/resources/actions/favorita/ObtainSingleFavoritaAction.java
@@ -1,8 +1,8 @@
-package com.sloy.sevibus.resources.actions;
+package com.sloy.sevibus.resources.actions.favorita;
 
 import com.google.common.base.Optional;
 import com.sloy.sevibus.model.tussam.Favorita;
-import com.sloy.sevibus.resources.datasource.FavoritaDataSource;
+import com.sloy.sevibus.resources.datasource.favorita.FavoritaDataSource;
 
 import rx.Observable;
