+++ SeviBus/src/main/java/com/sloy/sevibus/ui/fragments/FavoritasListFragment.java
@@ -16,8 +16,8 @@ import com.sloy.sevibus.model.tussam.Favorita;
 import com.sloy.sevibus.model.tussam.Linea;
 import com.sloy.sevibus.resources.Debug;
 import com.sloy.sevibus.resources.StuffProvider;
-import com.sloy.sevibus.resources.actions.ObtainFavoritasAction;
-import com.sloy.sevibus.resources.actions.ReorderFavoritasAction;
+import com.sloy.sevibus.resources.actions.favorita.ObtainFavoritasAction;
+import com.sloy.sevibus.resources.actions.favorita.ReorderFavoritasAction;
 import com.sloy.sevibus.ui.activities.ParadaInfoActivity;
 import com.sloy.sevibus.ui.adapters.DragFavoritaCallback;
 import com.sloy.sevibus.ui.adapters.FavoritasAdapter;
