+++ SeviBus/src/main/java/com/sloy/sevibus/ui/fragments/ParadaInfoFragment.java
@@ -43,10 +43,10 @@ import com.sloy.sevibus.resources.AnalyticsTracker;
 import com.sloy.sevibus.resources.Debug;
 import com.sloy.sevibus.resources.StuffProvider;
 import com.sloy.sevibus.resources.TimeTracker;
-import com.sloy.sevibus.resources.actions.DeleteFavoritaAction;
-import com.sloy.sevibus.resources.actions.ObtainLlegadasAction;
-import com.sloy.sevibus.resources.actions.ObtainSingleFavoritaAction;
-import com.sloy.sevibus.resources.actions.SaveFavoritaAction;
+import com.sloy.sevibus.resources.actions.favorita.DeleteFavoritaAction;
+import com.sloy.sevibus.resources.actions.llegada.ObtainLlegadasAction;
+import com.sloy.sevibus.resources.actions.favorita.ObtainSingleFavoritaAction;
+import com.sloy.sevibus.resources.actions.favorita.SaveFavoritaAction;
 import com.sloy.sevibus.ui.activities.BaseActivity;
 import com.sloy.sevibus.ui.activities.PreferenciasActivity;
 import com.sloy.sevibus.ui.widgets.LlegadasList;
