+++ SeviBus/src/test/java/com/sloy/sevibus/ui/mvp/presenter/FavoritasMainPresenterTest.java
@@ -2,7 +2,7 @@ package com.sloy.sevibus.ui.mvp.presenter;
 
 import com.sloy.sevibus.model.tussam.Favorita;
 import com.sloy.sevibus.model.tussam.Parada;
-import com.sloy.sevibus.resources.actions.ObtainFavoritasAction;
+import com.sloy.sevibus.resources.actions.favorita.ObtainFavoritasAction;
 
 import org.junit.Before;
 import org.junit.Test;
